chore: add integration test and make them work
