+++ test/testdata/backup/statefulset_pod.yaml
@@ -18,9 +18,9 @@ spec:
             fieldRef:
               apiVersion: v1
               fieldPath: metadata.name
-        - name: KB_REPLICASETS_PRIMARY_N
+        - name: KB_REPLICASETS_N
           value: '1'
-        - name: KB_REPLICASETS_PRIMARY_0_HOSTNAME
+        - name: KB_REPLICASETS_0_HOSTNAME
           value: wesql-cluster-replicasets-primary-0
       image: 'docker.io/apecloud/apecloud-mysql-server:latest'
       imagePullPolicy: IfNotPresent
