+++ test/testdata/cli_testdata/mysql_cd.yaml
@@ -42,11 +42,11 @@ spec:
             args:
               - >
                 cluster_info="";
-                for (( i=0; i<$KB_REPLICASETS_PRIMARY_N; i++ )); do
+                for (( i=0; i<$KB_REPLICASETS_N; i++ )); do
                   if [ $i -ne 0 ]; then
                     cluster_info="$cluster_info;";
                   fi;
-                  host=$(eval echo \$KB_REPLICASETS_PRIMARY_"$i"_HOSTNAME)
+                  host=$(eval echo \$KB_REPLICASETS_"$i"_HOSTNAME)
                   cluster_info="$cluster_info$host:13306";
                 done;
                 idx=0;
