+++ test/testdata/resources/mysql_cd.yaml
@@ -4,6 +4,9 @@ metadata:
   name: cluster-definition
 spec:
   type: state.mysql
+  connectionCredential:
+    username: "root"
+    password: "$(RANDOM_PASSWD)"
   components:
     - typeName: replicasets
       componentType: Stateful
@@ -36,11 +39,11 @@ spec:
             args:
               - >
                 cluster_info="";
-                for (( i=0; i<$KB_REPLICASETS_PRIMARY_N; i++ )); do
+                for (( i=0; i<$KB_REPLICASETS_N; i++ )); do
                   if [ $i -ne 0 ]; then
                     cluster_info="$cluster_info;";
                   fi;
-                  host=$(eval echo \$KB_REPLICASETS_PRIMARY_"$i"_HOSTNAME)
+                  host=$(eval echo \$KB_REPLICASETS_"$i"_HOSTNAME)
                   cluster_info="$cluster_info$host:13306";
                 done;
                 idx=0;
