--- test/testdata/resources/mysql_consensusset_cd.yaml
@@ -1,80 +1,0 @@
-apiVersion: dbaas.kubeblocks.io/v1alpha1
-kind: ClusterDefinition
-metadata:
-  name: cluster-definition-consensus
-spec:
-  type: state.mysql
-  components:
-    - typeName: replicasets
-      componentType: Consensus
-      consensusSpec:
-        leader:
-          name: "leader"
-          accessMode: ReadWrite
-        followers:
-          - name: "follower"
-            accessMode: Readonly
-        updateStrategy: BestEffortParallel
-      service:
-        ports:
-          - protocol: TCP
-            port: 3306
-      defaultReplicas: 3
-      podSpec:
-        containers:
-          - name: mysql
-            imagePullPolicy: IfNotPresent
-            ports:
-              - containerPort: 3306
-                protocol: TCP
-                name: mysql
-              - containerPort: 13306
-                protocol: TCP
-                name: paxos
-            env:
-              - name: MYSQL_ROOT_HOST
-                value: '%'
-              - name: MYSQL_ROOT_USER
-                value: root
-              - name: MYSQL_ROOT_PASSWORD
-              - name: MYSQL_ALLOW_EMPTY_PASSWORD
-                value: "yes"
-              - name: MYSQL_DATABASE
-                value: mydb
-              - name: MYSQL_USER
-                value: u1
-              - name: MYSQL_PASSWORD
-                value: u1
-              - name: CLUSTER_ID
-                value: "1"
-              - name: CLUSTER_START_INDEX
-                value: "1"
-              - name: REPLICATIONUSER
-                value: replicator
-              - name: REPLICATION_PASSWORD
-              - name: MYSQL_TEMPLATE_CONFIG
-              - name: MYSQL_CUSTOM_CONFIG
-              - name: MYSQL_DYNAMIC_CONFIG
-            command: ["/bin/bash", "-c"]
-            args:
-              - >
-                cluster_info="";
-                for (( i=0; i<$KB_REPLICASETS_N; i++ )); do
-                  if [ $i -ne 0 ]; then
-                    cluster_info="$cluster_info;";
-                  fi;
-                  host=$(eval echo \$KB_REPLICASETS_"$i"_HOSTNAME)
-                  cluster_info="$cluster_info$host:13306";
-                done;
-                idx=0;
-                while IFS='-' read -ra ADDR; do
-                  for i in "${ADDR[@]}"; do
-                    idx=$i;
-                  done;
-                done <<< "$KB_POD_NAME";
-                echo $idx;
-                cluster_info="$cluster_info@$(($idx+1))";
-                echo $cluster_info;
-                docker-entrypoint.sh mysqld --cluster-start-index=1 --cluster-info="$cluster_info" --cluster-id=1
-  connectionCredential:
-    user: root
