--- test/testdata/resources/mysql_consensusset_cv.yaml
@@ -1,13 +1,0 @@
-apiVersion: dbaas.kubeblocks.io/v1alpha1
-kind: ClusterVersion
-metadata:
-  name: cluster-version-consensus
-spec:
-  clusterDefinitionRef: cluster-definition-consensus
-  components:
-    - type: replicasets
-      podSpec:
-        containers:
-          - name: mysql
-            image: docker.io/apecloud/wesql-server:latest
-            imagePullPolicy: IfNotPresent
