--- test/testdata/resources/mysql_consensusset_pod.yaml
@@ -1,66 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  labels:
-    controller-revision-hash: mock-version
-  name: my-name
-  namespace: default
-spec:
-  containers:
-    - args:
-      command:
-        - /bin/bash
-        - -c
-      env:
-        - name: KB_POD_NAME
-          valueFrom:
-            fieldRef:
-              apiVersion: v1
-              fieldPath: metadata.name
-        - name: KB_REPLICASETS_N
-          value: "3"
-        - name: KB_REPLICASETS_0_HOSTNAME
-          value: clusterepuglf-wesql-test-0
-        - name: KB_REPLICASETS_1_HOSTNAME
-          value: clusterepuglf-wesql-test-1
-        - name: KB_REPLICASETS_2_HOSTNAME
-          value: clusterepuglf-wesql-test-2
-      image: docker.io/apecloud/wesql-server:latest
-      imagePullPolicy: IfNotPresent
-      name: mysql
-      ports:
-        - containerPort: 3306
-          name: mysql
-          protocol: TCP
-        - containerPort: 13306
-          name: paxos
-          protocol: TCP
-      volumeMounts:
-        - mountPath: /var/run/secrets/kubernetes.io/serviceaccount
-          name: kube-api-access-2rhsb
-          readOnly: true
-  dnsPolicy: ClusterFirst
-  enableServiceLinks: true
-  restartPolicy: Always
-  serviceAccount: default
-  serviceAccountName: default
-
-  volumes:
-    - name: kube-api-access-2rhsb
-      projected:
-        defaultMode: 420
-        sources:
-          - serviceAccountToken:
-              expirationSeconds: 3607
-              path: token
-          - configMap:
-              items:
-                - key: ca.crt
-                  path: ca.crt
-              name: kube-root-ca.crt
-          - downwardAPI:
-              items:
-                - fieldRef:
-                    apiVersion: v1
-                    fieldPath: metadata.namespace
-                  path: namespace
